try LogicFrame Scale(0.01f)

+++ idleshare/gamelib/src/hundun/gdxgame/idleshare/gamelib/export/IIdleFrontend.java
@@ -4,5 +4,5 @@ public interface IIdleFrontend {
     int getLogicFramePerSecond();
 
     boolean modLogicFrameSecondZero(int mod);
-    int getSecond();
+    float getSecond();
 }
